requirements.txt
biopy-isatab==0.1.1
Django==1.9.7
django-tables2==1.2.2
django-autocomplete-light==2.3.3
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0
djangorestframework-csv==1.4.1
django_rest_swagger==0.3.7
pandas==0.18.1
scipy
numpy
django-widget-tweaks==1.4.1
